requirements.txt
RPi.GPIO
pytest
smbus2